# Conflicts:
#	HNCHOME/Program.cs

+++ HNCHOME/Areas/Admin/Controllers/CustomerController.cs
@@ -25,19 +25,26 @@ namespace HNCHOME.Areas.Admin.Controllers
         {
             try
             {
+                if(customer == null)
+                {
+                    throw new Exception();
+                }
                 _customerRepository.Insert(customer);
                 return Ok(new { Code = 200, Res = "Successfully" });
             }
             catch (Exception)
             {
                 return BadRequest(new { Code = 500, Res = "Unsuccessfully" });
-                throw;
             }
         }
         public IActionResult DeleteCustomer(Guid id)
         {
             try
             {
+                if(id == Guid.Empty)
+                {
+                    throw new Exception();
+                }
                 _customerRepository.Delete(id);
                 return Ok(new {Res= "Successfully" });
             }
@@ -49,7 +56,18 @@ namespace HNCHOME.Areas.Admin.Controllers
         [HttpGet]
         public void GetCustomerDetail(Guid id)
         {
-            ViewBag.customer=_customerRepository.GetById(id);
+            try
+            {
+                if(id== Guid.Empty)
+                {
+                    throw new Exception();
+                }
+                ViewBag.customer = _customerRepository.GetById(id);
+            }
+            catch (Exception e)
+            {
+                throw;
+            }
         }
     }
 }
